refactor(Validator): Validator클래스 분리, 패키지 구조 정리
1. NameValidator 정의
2. RoundValidator 정의
3. utils패키지에 Const, NameValidator, RoundValidator 위치
4. domain패키지에 Car, Race 위치
